5_task/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0